addDragEvents
created file Game-dragEvents.js which has two methods: addDragEvents - which handles tiles that are stored in the .stand and player wants to move them to .board.
and another method: moveTilesAroundBoard which handles tiles that been placed on the board. if user hasnt clicked 'skip round', we can still move them around the board, or put back to stand.