changes
first i took away export default class Game, in the end of the code i added 3 rows that will do that, plus are going to import Game-dropEvents.js and its methods.
added method changePossibleToMoveToFalse, which loops through all tiles on the board and sets tile.possibleToMove = false.
calling this method when user skips round.
now in render i added some more data into tile, such as data-player (to check whos tile is that), data-tile: index of the tile on board(0-224), and data-possibletomove (it simply shows or not). also added span with tile.points.